se agrega Test en Back

+++ QuickParked-Back/src/main/java/com/example/quickparkedback/Repository/ISlotParkingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ISlotParkingRepository extends JpaRepository<Slotparking,Integer> {
+public interface ISlotParkingRepository extends JpaRepository<Slotparking,Long> {
 }
