+++ QuickParked-Back/src/main/java/com/example/quickparkedback/Repository/ITypeVehicleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ITypeVehicleRepository extends JpaRepository<Typevehicle,Integer> {
+public interface ITypeVehicleRepository extends JpaRepository<Typevehicle,Long> {
 }
