+++ QuickParked-Back/src/main/java/com/example/quickparkedback/Repository/IVehicleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IVehicleRepository extends JpaRepository <Vehicle,Integer> {
+public interface IVehicleRepository extends JpaRepository <Vehicle,Long> {
 }
